requirements.txt
matplotlib>=3.9.0
numpy>=1.26.4
pyyaml>=6.0
adjustText>=1.2.0
pandas>=2.0
openpyxl>=3.1.0
plotly>=5.0.0
